feat: project reorganization by adding lading page
In order to improve the application's architecture, the project in
Next.js has been modified in order to implement a very basic
mono-repository approach. Nonetheless, it is still a single package. The
idea is to take advantage of Next.js tools such as code split or
tree-shaking in order to reduce bundle size sent to the user.